refactor: move useMultiStageQueryLoadingState to @/hooks
