+++ src/pages/models/index.tsx
@@ -3,7 +3,7 @@ import { FC, ReactElement } from "react";
 import { PageBase, PageContentContainer } from "@/components/layouts";
 import { ModelsTable, PageTitle } from "@/components/ui/";
 import { useModelsWithInstances } from "@/services/model";
-import { useMultiStageQueryLoadingState } from "@/services/useMultiStageQueryLoadingState";
+import { useMultiStageQueryLoadingState } from "@/hooks/useMultiStageQueryLoadingState";
 
 interface GetLayOutProps {
   page: ReactElement;
